Refactor: 공간 관리 컨트롤러 분리
